+++ src/main/kotlin/com/teame1i4/newsfeed/domain/post/service/PostService.kt
@@ -2,7 +2,9 @@ package com.teame1i4.newsfeed.domain.post.service
 
 import com.teame1i4.newsfeed.domain.exception.ModelNotFoundException
 import com.teame1i4.newsfeed.domain.exception.TypeNotFoundException
+import com.teame1i4.newsfeed.domain.exception.UnauthorizedAccessException
 import com.teame1i4.newsfeed.domain.exception.YouTubeUrlNotValidException
+import com.teame1i4.newsfeed.domain.member.adapter.MemberDetails
 import com.teame1i4.newsfeed.domain.post.dto.CreatePostRequest
 import com.teame1i4.newsfeed.domain.post.dto.PostResponse
 import com.teame1i4.newsfeed.domain.post.dto.PostWithCommentResponse
@@ -14,6 +16,7 @@ import com.teame1i4.newsfeed.domain.musictype.repository.MusicTypeRepository
 import com.teame1i4.newsfeed.domain.post.model.*
 import jakarta.transaction.Transactional
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.stereotype.Service
 
 @Service
@@ -23,17 +26,17 @@ class PostService(
     private val memberRepository: MemberRepository,
 ) {
 
-    fun createPost(request: CreatePostRequest): PostResponse {
-        musicTypeRepository.findByIdOrNull(request.musicType) ?: throw TypeNotFoundException(request.musicType)
-        memberRepository.findByIdOrNull(request.memberId) ?: throw ModelNotFoundException("member", request.memberId)
-
+    @PreAuthorize("hasRole('USER')")
+    fun createPost(request: CreatePostRequest, member: MemberDetails): PostResponse {
+        if(!musicTypeRepository.existsById(request.musicType)) throw TypeNotFoundException(request.musicType)
+//        memberRepository.findByIdOrNull(member.memberId) ?: throw ModelNotFoundException("member", member.memberId)
         val youtubeId = extractYoutubeId(request.musicUrl)
 
         val post = Post(
             title = request.title,
             content = request.content,
             musicUrl = youtubeId,
-            memberId = request.memberId,
+            memberId = member.memberId,
             musicType = request.musicType,
             tags = "#" + request.tags.joinToString("#") + "#"
         )
@@ -43,12 +46,16 @@ class PostService(
         musicType.updateCountPost(true)
         musicTypeRepository.save(musicType)
 
-        return postRepository.save(post).toResponse()
+        return postRepository.save(post).toResponse(memberRepository.findByIdOrNull(post.memberId)!!)
     }
 
+    @PreAuthorize("hasRole('USER')")
     @Transactional
-    fun deletePost(postId: Long) {
+    fun deletePost(postId: Long, member: MemberDetails) {
         val post: Post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+
+        if(post.memberId != member.memberId) throw UnauthorizedAccessException()
+
         postRepository.delete(post)
 
         val musicType: MusicType =
@@ -57,9 +64,11 @@ class PostService(
         musicTypeRepository.save(musicType)
     }
 
+    @PreAuthorize("hasRole('USER')")
     @Transactional
-    fun updatePost(postId: Long, request: UpdatePostRequest): PostResponse {
+    fun updatePost(postId: Long, request: UpdatePostRequest, member: MemberDetails): PostResponse {
         val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        if(post.memberId != member.memberId) throw UnauthorizedAccessException()
         var musicType: MusicType =
             musicTypeRepository.findByIdOrNull(request.musicType) ?: throw TypeNotFoundException(request.musicType)
 
@@ -74,7 +83,7 @@ class PostService(
         musicType.updateCountPost(true)
         musicTypeRepository.save(musicType)
 
-        return postRepository.save(post).toResponse()
+        return postRepository.save(post).toResponse(memberRepository.findByIdOrNull(post.memberId)!!)
     }
 
     fun getPosts(tag: String?, title: String?, musicType: String?, memberId: Long?): List<PostResponse> {
@@ -95,7 +104,7 @@ class PostService(
             )
             else postRepository.findAllByPostStatusOrderByCreatedAtDesc(PostStatus.PUBLIC)
 
-        return posts.map { it.toResponse() }
+        return posts.map { it.toResponse(memberRepository.findByIdOrNull(it.memberId)!!) }
     }
 
     @Transactional
@@ -105,7 +114,7 @@ class PostService(
 
         post.comments.sortBy { it.createdAt }
 
-        return post.toWithCommentResponse()
+        return post.toWithCommentResponse(memberRepository.findByIdOrNull(post.memberId)!!)
     }
 
     private fun extractYoutubeId(musicUrl: String): String {
